use f64

+++ rust/src/classification/logistic_regression.rs
@@ -3,13 +3,13 @@ use crate::common::functions::sigmoid;
 use na::{DMatrix, DVector};
 // use std::collections::HashSet;
 pub struct LogisticRegressionResult {
-    pub weights: DMatrix<f32>,
+    pub weights: DMatrix<f64>,
     pub n_features: usize,
 }
 
 #[no_mangle]
 pub unsafe extern "C" fn logistic_regression(
-    x_ptr: *const f32,
+    x_ptr: *const f64,
     y_ptr: *const u32,
     x_len: usize,
     y_len: usize,
@@ -23,7 +23,7 @@ pub unsafe extern "C" fn logistic_regression(
     // let classes: HashSet<&u32> = HashSet::from_iter(y.into_iter());
     // let n_classes = classes.len();
 
-    let mut weights: DMatrix<f32> = DMatrix::zeros(1, n_features);
+    let mut weights: DMatrix<f64> = DMatrix::zeros(1, n_features);
     for i in 0..n_features {
         weights[i] = 1.0;
     }
@@ -31,30 +31,30 @@ pub unsafe extern "C" fn logistic_regression(
     let mut data = DMatrix::from_row_slice(x_len, n_features, x);
 
     for i in 0..epochs {
-        let hypotheses: Vec<f32> = data
+        let hypotheses: Vec<f64> = data
             .row_iter_mut()
             .map(|x| sigmoid(weights.dot(&x)))
             .collect();
         if i % 100 == 0 && !silent {
-            let error: f32 = (0..x_len)
+            let error: f64 = (0..x_len)
                 .map(|i| {
                     let h_i = hypotheses.get(i).unwrap();
-                    let y_i = y[i] as f32;
+                    let y_i = y[i] as f64;
 
-                    return (y_i * h_i.log2() + (1.0 - y_i) * (1.0 - h_i).log2()) as f32;
+                    return (y_i * h_i.log2() + (1.0 - y_i) * (1.0 - h_i).log2()) as f64;
                 })
-                .sum::<f32>()
-                / x_len as f32;
+                .sum::<f64>()
+                / x_len as f64;
             println!("Epoch <{}: Current Errors {}", i, error);
         }
 
-        let errors: Vec<f32> = (0..x_len)
-            .map(|i| hypotheses.get(i).unwrap() - (y[i] as f32))
+        let errors: Vec<f64> = (0..x_len)
+            .map(|i| hypotheses.get(i).unwrap() - (y[i] as f64))
             .collect();
 
         let err = DVector::from_vec(errors);
 
-        let weight_updates: Vec<f32> = (0..n_features)
+        let weight_updates: Vec<f64> = (0..n_features)
             .map(|i| {
                 let x_i = data.column(i);
                 0.001 * x_i.dot(&err)
@@ -78,8 +78,8 @@ pub unsafe extern "C" fn logistic_regression(
 #[no_mangle]
 pub unsafe extern "C" fn logistic_regression_predict_y(
     res: *const LogisticRegressionResult,
-    x_ptr: *const f32,
-) -> f32 {
+    x_ptr: *const f64,
+) -> f64 {
     let _res = &*res;
 
     let x = std::slice::from_raw_parts(x_ptr, _res.n_features);
